DISCO 3756 - Fix a circular import issue in domain_category_mapping
