+++ merino/jobs/navigational_suggestions/__init__.py
@@ -16,7 +16,7 @@ from merino.configs import settings as config
 from merino.jobs.navigational_suggestions.partner_favicons import PARTNER_FAVICONS
 from merino.jobs.navigational_suggestions.utils import AsyncFaviconDownloader
 from merino.utils.gcs.gcs_uploader import GcsUploader
-from merino.jobs.utils.domain_category_mapping import DOMAIN_MAPPING
+from merino.utils.domain_categories.domain_category_mapping import DOMAIN_MAPPING
 from merino.jobs.navigational_suggestions.domain_data_downloader import (
     DomainDataDownloader,
 )
@@ -27,7 +27,7 @@ from merino.jobs.navigational_suggestions.domain_metadata_extractor import (
 from merino.jobs.navigational_suggestions.domain_metadata_uploader import (
     DomainMetadataUploader,
 )
-from merino.providers.suggest.base import Category
+from merino.utils.domain_categories.models import Category
 from merino.utils.blocklists import TOP_PICKS_BLOCKLIST
 from merino.jobs.navigational_suggestions.custom_favicons import get_custom_favicon_url
