+++ merino/utils/domain_categories/__init__.py
@@ -1,0 +1,1 @@
+"""Module for domain to SERP categories."""
