+++ tests/integration/api/v1/suggest/test_suggest_top_picks.py
@@ -13,10 +13,10 @@ from fastapi.testclient import TestClient
 from pydantic import HttpUrl
 
 from merino.configs import settings
-from merino.providers.suggest.base import Category
 from merino.providers.suggest.top_picks.backends.top_picks import TopPicksBackend
 from merino.providers.suggest.top_picks.provider import Provider, Suggestion
 from tests.integration.api.v1.types import Providers
+from merino.utils.domain_categories.models import Category
 
 
 @pytest.fixture(name="domain_blocklist")
